refactor(images carousel): use useStaticContent
- moved fetchpolicy and refreshTime from ConnectedImagesCarousel to ImagesCarousel
- used useStaticContent in ConnectedImagesCarousel

SVA-150
